@@ -1,16 +1,17 @@
 {
   "name": "conseiljs",
-  "version": "0.1.39",
+  "version": "0.1.40",
   "description": "Client-side library for dApp development.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "test": "mocha",
-    "integration-test": "./node_modules/mocha/bin/mocha --timeout 1000000 -r ts-node/register integration_test/*.spec.ts",
+    "test": "mocha test/*.spec.ts",
+    "integration-test": "mocha --compilers ts-node/register --require source-map-support/register --require ts-node/register --full-trace --timeout 10000 integration_test/*.spec.ts",
     "format": "eslint --fix-dry-run src/*/*.ts",
     "package": "rm -rf ./dist && tsc",
     "build": "webpack && rm -rf ./build/dist",
-    "coverage": "nyc mocha || nyc report --reporter=text-lcov | coveralls && rm -rf ./coverage"
+    "coverage": "nyc mocha || nyc report --reporter=text-lcov | coveralls && rm -rf ./coverage",
+    "doc": "rm -rf ./doc && typedoc --out ./doc ./src"
   },
   "repository": {
     "type": "git",
@@ -18,7 +19,10 @@
   },
   "keywords": [
     "tezos",
-    "conseil"
+    "conseil",
+    "blockchain analytics",
+    "blockchain reporting",
+    "blockchain data access"
   ],
   "author": "Cryptonomic Inc",
   "license": "GPL-3.0",
@@ -45,10 +49,9 @@
     "instrument": true
   },
   "dependencies": {
-    "@ledgerhq/hw-transport": "4.21.0",
-    "@ledgerhq/hw-transport-node-hid": "^4.22.0",
-    "awesome-typescript-loader": "^5.0.0",
-    "aws-sdk": "^2.326.0",
+    "@ledgerhq/hw-transport": "4.24.0",
+    "@ledgerhq/hw-transport-node-hid": "4.24.0",
+    "aws-sdk": "^2.395.0",
     "babel-runtime": "6.26.0",
     "base-n": "git://github.com/yourcodesucks/base-n.git",
     "basil-tezos-ledger": "0.0.1",
@@ -58,36 +61,39 @@
     "https": "^1.0.0",
     "libsodium-wrappers": "^0.5.4",
     "libsodium-wrappers-sumo": "^0.5.4",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
+    "node-fetch": "^2.3.0",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
     "@types/chai": "^4.1.3",
     "@types/mocha": "^5.2.0",
-    "@types/node": "^10.3.0",
-    "@types/node-fetch": "^2.1.1",
+    "@types/node": "^10.12.19",
+    "@types/node-fetch": "^2.1.4",
     "@types/zxcvbn": "^4.4.0",
+    "awesome-typescript-loader": "^5.0.0",
     "chai": "^4.1.2",
     "coveralls": "^3.0.2",
     "eslint": "^5.12.0",
     "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-import": "^2.16.0",
     "eslint-plugin-typescript": "^0.14.0",
     "mocha": "^5.2.0",
-    "node-fetch": "^2.1.1",
     "nyc": "^13.1.0",
-    "source-map-support": "^0.5.9",
+    "source-map-loader": "^0.2.4",
+    "source-map-support": "^0.5.10",
     "ts-loader": "^4.3.1",
     "ts-node": "^6.0.5",
+    "tsconfig-paths-webpack-plugin": "^3.2.0",
     "typedoc": "^0.11.1",
     "typescript": "^2.8.3",
     "typescript-eslint-parser": "^21.0.2",
     "uglifyjs-webpack-plugin": "^1.2.5",
-    "webpack": "^4.28.1",
-    "webpack-cli": "^3.0.3"
+    "webpack": "^4.29.0",
+    "webpack-cli": "^3.2.1",
+    "webpack-node-externals": "^1.7.2"
   },
   "peerDependencies": {
-    "node-fetch": "^2.1.1"
+    "node-fetch": "^2.3.0"
   },
   "eslintConfig": {
     "parser": "typescript-eslint-parser",
@@ -97,12 +103,32 @@
     },
     "extends": "airbnb-base",
     "rules": {
-      "comma-dangle": [2, "never"],
-      "curly": [2, "all"],
-      "indent": [1, 4],
-      "max-len": [2, { "code": 200, "ignoreTrailingComments": true }],
-      "no-tabs": [2],
-      "quotes": [2, "double"]
+      "comma-dangle": [
+        2,
+        "never"
+      ],
+      "curly": [
+        2,
+        "all"
+      ],
+      "indent": [
+        1,
+        4
+      ],
+      "max-len": [
+        2,
+        {
+          "code": 200,
+          "ignoreTrailingComments": true
+        }
+      ],
+      "no-tabs": [
+        2
+      ],
+      "quotes": [
+        2,
+        "double"
+      ]
     }
   }
 }
