@@ -10,6 +10,7 @@ describe('Block fetchers', () => {
     it('should correctly fetch blocks', async () => {
         const head = await TezosConseilQuery.getBlockHead(conseilURL, conseilApiKey);
         expect(head.hash.startsWith('B')).to.equal(true);
+
         const aBlock = await TezosConseilQuery.getBlock(conseilURL, head.hash, conseilApiKey);
         expect(aBlock.block.hash).to.equal(head.hash);
     });
@@ -20,10 +21,13 @@ describe('Operation fetchers', () => {
         const emptyFilter = TezosConseilQuery.getEmptyTezosFilter();
         const opFilter = {...emptyFilter, limit: 10, operation_kind: ['transaction']};
         const ops = await TezosConseilQuery.getOperations(conseilURL, opFilter, conseilApiKey);
+
         expect(ops.length).to.equal(10);
+
         const opGroup = await TezosConseilQuery.getOperationGroup(conseilURL, ops[0].operationGroupHash, conseilApiKey);
         expect(opGroup.operation_group.hash).to.equal(ops[0].operationGroupHash);
-        const opGroupsFilter = {...emptyFilter, limit: 10, operation_id: [opGroup.operation_group.hash]};
+
+        const opGroupsFilter = {...emptyFilter, limit: 10, operation_id:  [opGroup.operation_group.hash]};
         const opGroups = await TezosConseilQuery.getOperationGroups(conseilURL, opGroupsFilter, conseilApiKey);
         expect(opGroups.length).to.equal(1)
     });
