@@ -82,5 +82,29 @@ export interface Operation {
     delegatable?: boolean,
     destination?: string,
     amount?: string,
-    script?: string
+    script?: string | object;
+}
+
+export interface ContractOriginationOperation extends Operation {
+    script: object;
+}
+
+export interface ContractInvocationOperation extends Operation {
+    parameters: object;
+}
+
+/**
+ * Output of operation signing.
+ */
+export interface SignedOperationGroup {
+    bytes: Buffer;
+    signature: string;
+}
+
+/**
+ * Result of a successfully sent operation
+ */
+export interface OperationResult {
+    results: AlphaOperationsWithMetadata;
+    operationGroupID: string;
 }
